Split rules into per-library modules, remove dead code, update README
Phase 1 refactoring:
- Split rules_libraries.py (574 lines, 18 rules, 10 libraries) into
  individual modules under rules/ package: django.py, fastapi.py,
  sqlalchemy.py, flask.py, celery.py, pandas.py, requests_rules.py,
  pydantic.py, pytest_rules.py, drf.py
- Moved architecture.py and py314.py into rules/ package
- rules/__init__.py assembles ALL_RULES from all modules
- pack.py simplified to single import

Dead code removal:
- Engine._config (assigned, never used)
- PythonContext.tables (alias for models, zero callers)
- ModelInfo.table_name (declared, never assigned or read)

README rewritten for Python-only focus with full rule category and
library-specific rule tables.

Closes #4, partially addresses #16, #17, #18
