+++ src/gaudi/packs/python/rules/architecture.py
