+++ src/gaudi/packs/python/rules/drf.py
@@ -1,0 +1,57 @@
+# ABOUTME: Django REST Framework architectural rules for Gaudí Python pack.
+# ABOUTME: Covers missing permission classes and throttling on ViewSets.
+from __future__ import annotations
+
+from gaudi.core import Rule, Finding, Severity, Category
+from gaudi.packs.python.context import PythonContext
+
+
+class DRFNoPermissionClass(Rule):
+    code = "DRF-SEC-001"
+    severity = Severity.WARN
+    category = Category.SECURITY
+    message_template = "ViewSet without explicit permission_classes in '{file}'"
+    recommendation_template = (
+        "Set permission_classes on every ViewSet. Relying on DEFAULT_PERMISSION_CLASSES "
+        "is fragile — a settings change could expose all endpoints."
+    )
+
+    def check(self, context: PythonContext) -> list[Finding]:
+        findings = []
+        for f in context.files:
+            if "rest_framework" not in str(f.imports):
+                continue
+            try:
+                source = f.path.read_text(encoding="utf-8", errors="replace")
+            except Exception:
+                continue
+            if ("ViewSet" in source or "APIView" in source) and "permission_classes" not in source:
+                findings.append(self.finding(file=f.relative_path))
+        return findings
+
+
+class DRFNoThrottling(Rule):
+    code = "DRF-SCALE-001"
+    severity = Severity.INFO
+    category = Category.SCALABILITY
+    message_template = "API view without throttle_classes in '{file}'"
+    recommendation_template = (
+        "Add throttle_classes to prevent abuse. Public APIs without rate limiting "
+        "are vulnerable to DDoS and abuse."
+    )
+
+    def check(self, context: PythonContext) -> list[Finding]:
+        findings = []
+        for f in context.files:
+            if "rest_framework" not in str(f.imports):
+                continue
+            try:
+                source = f.path.read_text(encoding="utf-8", errors="replace")
+            except Exception:
+                continue
+            if ("ViewSet" in source or "APIView" in source) and "throttle_classes" not in source:
+                findings.append(self.finding(file=f.relative_path))
+        return findings
+
+
+DRF_RULES = [DRFNoPermissionClass(), DRFNoThrottling()]
